backend/requirements.txt
Flask==2.1.1
Werkzeug==2.1.2
Flask-Cors==3.0.10
PyJWT==2.3.0